fix(94): Change story file extension from ts to tsx to fix issue with rendering the ExploreButton component within a router decorator
